restricting non user from updating the room

+++ base/views.py
@@ -1,7 +1,10 @@
+from django.http.response import HttpResponse
+from django.http import HttpResponse
 from django.shortcuts import render,redirect
 from django.contrib import messages
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate,login,logout
 from .models import Room,Topic
 from .forms import RoomForm
@@ -55,6 +58,8 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'base/room.html',context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -66,9 +71,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html',context)
 
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed !')
 
     if request.method == 'POST':
         form = RoomForm(request.POST,instance = room)
@@ -79,8 +88,14 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)    
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed !')
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
